Move leaseDefaults to /util
