+++ frontend/app/api/retrieve/lease-settings/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { PrismaClient, Role } from "@prisma/client";
 import { requireRole } from "../../../../services/auth";
-import { defaultLeaseSettings } from "../../../../services/leaseDefaults";
+import { defaultLeaseSettings } from "../../../../util/leaseDefaults";
 import type { IRoomRate } from "../../../../util/models";
 
 const prisma = new PrismaClient();
